+++ src/main/java/modelos/TransferenciaRealizada.java
@@ -16,7 +16,7 @@ public class TransferenciaRealizada extends Movimiento{
 	@ManyToOne
 	private Cuenta cuentaDestino;
 
-	public TransferenciaRealizada(LocalDate fechayHora, float monto, String descripcion, Cuenta cuenta, Cuenta cuentaDestino) {
+	public TransferenciaRealizada(LocalDate fechayHora, Double monto, String descripcion, Cuenta cuenta, Cuenta cuentaDestino) {
 		super(fechayHora, monto, descripcion, cuenta);
 	
 		this.cuentaDestino=cuentaDestino;
